refactor(examples): split snapshot.rs into main/scenes/encode (move-only)
Issue #481: the 2251-LOC single-file example was a mode-dispatch main() mixed
with scene builders + PNG/GIF encoding. Split into a directory example
`examples/snapshot/` (name unchanged, so `cargo run --example snapshot` and the
gen-media/just recipes are unaffected — only the Cargo.toml [[example]] path
moved):
- main.rs — args (SnapshotArgs/clap), shared consts, default_projects_root,
  parse_navigations, due_navigations, main() dispatch, the cross-module tests.
- scenes.rs — the SceneState/fixture builders (sample/meeting/dashboard/anim/
  openclaw/live-capture + fill_sample_agents).
- encode.rs — PNG/GIF encode + crop math + the walkable-overlay debug painter.

Pure relocation: every fn/struct/const body moved VERBATIM; only module
structure, `pub(crate)` on the 12 cross-module items, and the split `use` blocks
changed (+50 net structural lines, no body edits). Docs-currency: the 4 live
CLAUDE.md/offscreen.rs prose refs to `examples/snapshot.rs` repointed to
`examples/snapshot/` (frozen REVIEW-LEDGER left as-is).

Byte-identical output verified: md5 vs an origin/main build across 15 render
modes (11 static/overlay incl. the debug-walkable BFS trio + 4 GIF paths) — all
IDENTICAL. Build + clippy(-D warnings) clean; 17 example tests pass.
