+++ crates/pixtuoid/src/tui/CLAUDE.md
@@ -97,7 +97,7 @@ tui/
 
 ## Where to look
 
-- "How is the office rendered (pixel pass → terminal)?" → the world render is `pixtuoid_scene::pixel_painter::render_to_rgb_buffer` (see [`../../../pixtuoid-scene/CLAUDE.md`](../../../pixtuoid-scene/CLAUDE.md) for the pixel-pass split). The terminal-flush pass (`half-block + widgets + status footer`) is **`tui::renderer::draw_scene`**. `draw_scene` is the ONE consumer still on raw `render_to_rgb_buffer` (it needs the full `PixelPassResult` + immutable coffee borrows mid-flush); every headless surface goes through the shared `pixtuoid_scene::floor::render_floor` seam (#423) — the floor-slide calls it directly (projected scenes, caller-side eviction), while `floating::offscreen::OfficeRenderer` and the web hero each own a `pixtuoid_scene::floor::FloorSession` (the scene-owned session whose `render()` also runs the dual eviction); `TuiRenderer` composes the same session halves as `Vec<PerFloor>` + one `PerOffice`. `examples/snapshot.rs` saves the ratatui `TestBackend` via `save_backend_as_png` (a ▀-compressed PNG with footer).
+- "How is the office rendered (pixel pass → terminal)?" → the world render is `pixtuoid_scene::pixel_painter::render_to_rgb_buffer` (see [`../../../pixtuoid-scene/CLAUDE.md`](../../../pixtuoid-scene/CLAUDE.md) for the pixel-pass split). The terminal-flush pass (`half-block + widgets + status footer`) is **`tui::renderer::draw_scene`**. `draw_scene` is the ONE consumer still on raw `render_to_rgb_buffer` (it needs the full `PixelPassResult` + immutable coffee borrows mid-flush); every headless surface goes through the shared `pixtuoid_scene::floor::render_floor` seam (#423) — the floor-slide calls it directly (projected scenes, caller-side eviction), while `floating::offscreen::OfficeRenderer` and the web hero each own a `pixtuoid_scene::floor::FloorSession` (the scene-owned session whose `render()` also runs the dual eviction); `TuiRenderer` composes the same session halves as `Vec<PerFloor>` + one `PerOffice`. `examples/snapshot/` saves the ratatui `TestBackend` via `save_backend_as_png` (a ▀-compressed PNG with footer).
 - "How does the neon wall board work?" → the pixel-buffer half (dark panel + pulsing cyan border) is `pixtuoid_scene::pixel_painter::background::lighting.rs::paint_neon_panel`; `widgets/hud.rs::paint_wall_display` overlays 3 rows of ratatui text: **L1** brand + the `★ Star` CTA (right-flushed; its precise click target is `hud::star_hit_rect`, opening `REPO_URL`); **L2** the plain-English **mood pulse** (`board_mood_segments` — one `{glyph}{count} {word}` per non-zero present state, the ▲ "needs-you" beacon LEADING, NUMERIC + abbreviated to fit, from the SAME `scene_stats` the footer reads); **L3** office context (uptime `tooltip_dim` + floor breadcrumb + the `⬢gw` gateway chip). The board's cell-origin + WIDTH pin to the painted panel's dark INTERIOR — `hud::BOARD_W = pixtuoid_scene::pixel_painter::NEON_PANEL_INNER_W` (the outer `NEON_PANEL_W` minus its `NEON_PANEL_BORDER` on each side) and `hud::board_cell_origin` starts text at `NEON_PANEL_INNER_X` — so the lit text sits INSIDE the glowing frame (laying it to the full outer width overran it — the board-overflow fix; `paint_neon_panel` reads the SAME border const, so the dark cells it leaves == the cells text may fill). `board_cell_origin` is the ONE origin `paint_wall_display` AND `star_hit_rect` share. **The scrolling ticker was removed** (0.12.0) — `TickerQueue` + the `neon_ticker` theme color are gone; `marquee_window`/`MARQUEE_*` survive (dashboard/connection selected-row scroll).
 - "How does the theme PICKER UI work?" → the theme MODEL (the `Theme` struct, `ALL_THEMES`, `SourceColors`/`ApplianceColors`) lives in `pixtuoid_scene::theme` (see [`../../../pixtuoid-scene/CLAUDE.md`](../../../pixtuoid-scene/CLAUDE.md)). The UI is here: press `[t]` for a live preview picker (j/k or ↑↓ to navigate); each row shows a 2-cell swatch (`hud.rs::theme_swatch` → `neon_brand` + `carpet_base`) drawn from that row's own palette. The picker is dispatched in `tui/mod.rs`; `set_theme()` (on `TuiRenderer`) flushes the `FrameCache` so character recolors update immediately. Theme is saved on `[t]` picker Enter confirm in `tui/mod.rs`.
 - "How are popups framed?" → **every popup is borderless** (no outline) and shares ONE backing: `widgets/mod.rs::paint_card_backing(f, area, theme)` = a flat single-color drop **shadow** (a hand-rolled fg+bg dim — NOT ratatui's `Block::shadow`, which smears over half-blocks) + `Clear` + a solid `tooltip_bg` fill. **The shadow is a deliberately FLAT offset silhouette, not a soft penumbra** (`cast_drop_shadow`): the card's own rect is darkened by ONE uniform `SHADOW_FACTOR` (0.42) and offset `SHADOW_OFFSET` (1) cell down-right; the card paints over its own cells afterward, leaving an L-band (right column offset down, bottom row offset right, meeting at the corner) — all ONE color. The **bottom-most row is rendered TOP-HALF only** (`dim_cell`'s `top_half_only`: dims `fg`, leaves `bg` lit) so the bottom band reads as a 1px contact line instead of a full 2px cell, while the vertical right strip stays full cells. This is a pinned user preference (matches a supplied reference — flat single color, offset, darker, thin bottom); do NOT "restore" a gradient/falloff or per-band factor, and do NOT make the bottom a full 2px cell — the offset silhouette with the 1px top-half bottom is the intended look (`panel::borderless_panel_casts_a_flat_offset_shadow` pins it). Both card kinds delegate to it, so the look (incl. the shadow) has a single definition and can't drift: `tui/widgets/panel.rs::borderless_panel` calls it, then insets by a uniform `PANEL_PAD_X`/`PANEL_PAD_Y` (the breathing room that replaces the border — re-exported from `widgets`), draws an optional bold brand-colored title row, and returns the inner content `Rect` (Help / theme-picker / version-popup / dashboard / connection route through it, each `centered_in` adding `2×PANEL_PAD_*`); the hover/info tooltips (`tooltip.rs::framed_tooltip`, a `Padding::uniform(1)` text-only Block — the bg/clear/shadow now come from the backing) call it too. Chitchat speech bubbles are deliberately NOT this primitive (a bare 1-row `Paragraph`, no `Clear`/block) so they get no shadow. The title moved from the old border into the body. **The version popup's URL click-rect (`version_popup_url_rect`) derives its offsets from the SAME `PANEL_PAD_*` consts the painter insets by** (`url_x = popup_x + PANEL_PAD_X + URL_PREFIX.len()`, edges clipped at `±PANEL_PAD_*`) — keep them in lockstep (the phantom-browser-launch regression class).
